Move ErrorHandler and co to api-errors.go
