+++ tools/middlewares.go
@@ -323,42 +323,3 @@ func (c *TUMLiveContext) UserIsAdmin() bool {
 	}
 	return c.User.IsAdminOfCourse(*c.Course)
 }
-
-type RequestError struct {
-	Status        int
-	CustomMessage string
-	Err           error
-}
-
-func (r RequestError) Error() string {
-	if r.Err != nil {
-		return r.Err.Error()
-	} else {
-		return ""
-	}
-}
-
-func (r RequestError) ToResponse() gin.H {
-	res := gin.H{"status": r.Status, "message": r.CustomMessage}
-
-	if r.Err != nil {
-		res["error"] = r.Error()
-	}
-
-	return res
-}
-
-func ErrorHandler(c *gin.Context) {
-	c.Next()
-
-	if len(c.Errors) > 0 {
-		err := c.Errors[0]
-		switch err.Err.(type) {
-		case RequestError:
-			e := err.Err.(RequestError)
-			c.Errors = []*gin.Error{} // Clear errors so they don't get logged
-			c.JSON(e.Status, e.ToResponse())
-			return
-		}
-	}
-}
